网关请求超时配置

+++ src/Wing.GateWay/Config/Policy.cs
@@ -18,7 +18,7 @@
 
         public bool? UseJWTAuth { get; set; }
 
-        public int? HttpClientTimeOut { get; set; }
+        public double? HttpClientTimeOut { get; set; }
 
         public override bool Equals(object obj)
         {
